// src/app/map/_types/types.ts
export type datasType = {
  DTLSTATENM: string;
  BPLCNM: string;
  RDNWHLADDR: string;
  AR_CD_NAME: string;
  FT_KIND_NAME: string;
  FT_HOMEPAGE: string;
  FT_TITLE: string;
  FT_ADDR: string;
  FT_ADDR_DETAIL: string;
  FT_MONEY: string;
  FT_WD_TIME: string;
  FT_WE_TIME: string;
  FT_PHONE: string;
  FT_PARK: string;
};